+++ Package.swift
@@ -6,6 +6,9 @@ let package = Package(
     platforms: [
         .macOS(.v13)
     ],
+    dependencies: [
+        .package(url: "https://github.com/sparkle-project/Sparkle", exact: "2.9.3")
+    ],
     targets: [
         .target(
             name: "EverywhereCore",
@@ -15,7 +18,8 @@ let package = Package(
         ),
         .executableTarget(
             name: "Everywhere",
-            dependencies: ["EverywhereCore"]
+            dependencies: ["EverywhereCore", .product(name: "Sparkle", package: "Sparkle")],
+            linkerSettings: [.unsafeFlags(["-Xlinker", "-rpath", "-Xlinker", "@executable_path/../Frameworks"])]
         ),
         .testTarget(
             name: "EverywhereCoreTests",